renderer: micro optimizations

+++ src/device/gpu/render/texture_utils.h
@@ -6,21 +6,19 @@
 
 enum class ColorDepth { NONE, BIT_4, BIT_8, BIT_16 };
 
-INLINE uint16_t tex4bit(gpu::GPU* gpu, glm::ivec2 tex, glm::ivec2 texPage, glm::ivec2 clut) {
+// Using unsigned vectors allows compiler to generate slightly faster division code
+INLINE uint16_t tex4bit(gpu::GPU* gpu, glm::uvec2 tex, glm::uvec2 texPage, glm::uvec2 clut) {
     uint16_t index = gpuVRAM[texPage.y + tex.y][texPage.x + tex.x / 4];
-    uint16_t entry = (index >> ((tex.x & 3) * 4)) & 0xf;
+    uint8_t entry = (index >> ((tex.x & 3) * 4)) & 0xf;
     return gpuVRAM[clut.y][clut.x + entry];
 }
 
-INLINE uint16_t tex8bit(gpu::GPU* gpu, glm::ivec2 tex, glm::ivec2 texPage, glm::ivec2 clut) {
+INLINE uint16_t tex8bit(gpu::GPU* gpu, glm::uvec2 tex, glm::uvec2 texPage, glm::uvec2 clut) {
     uint16_t index = gpuVRAM[texPage.y + tex.y][texPage.x + tex.x / 2];
-    uint16_t entry = (index >> ((tex.x & 1) * 8)) & 0xff;
+    uint8_t entry = (index >> ((tex.x & 1) * 8)) & 0xff;
     return gpuVRAM[clut.y][clut.x + entry];
 }
 
-INLINE uint16_t tex16bit(gpu::GPU* gpu, glm::ivec2 tex, glm::ivec2 texPage) {
-    return gpuVRAM[texPage.y + tex.y][texPage.x + tex.x];
-    // TODO: In PSOne BIOS colors are swapped (r == b, g == g, b == r, k == k)
-}
+INLINE uint16_t tex16bit(gpu::GPU* gpu, glm::uvec2 tex, glm::uvec2 texPage) { return gpuVRAM[texPage.y + tex.y][texPage.x + tex.x]; }
 
 #undef gpuVRAM
